Bearing as integer.

+++ mls.h
@@ -41,7 +41,7 @@ class MLS {
     int   geoLocation();
     long  getMovement();
     float getDistance(float lat1, float long1, float lat2, float long2);
-    float getBearing(float lat1, float long1, float lat2, float long2);
+    int   getBearing(float lat1, float long1, float lat2, float long2);
     const char *getCardinal(float course);
     geo_t current;
     geo_t previous;
